fix UVData.get_antpairs printing and other tutorial issues for numpy 2.0

+++ pyuvdata/utils.py
@@ -805,9 +805,9 @@ def baseline_to_antnums(baseline, Nants_telescope):
         np.ascontiguousarray(baseline, dtype=np.uint64)
     )
     if return_array:
-        return ant1, ant2
+        return ant1.astype(int), ant2.astype(int)
     else:
-        return ant1.item(0), ant2.item(0)
+        return int(ant1.item(0)), int(ant2.item(0))
 
 
 def antnums_to_baseline(
